chore: split transaction file part one fields dedicated file
