+++ crates/blockifier/src/transaction/objects.rs
@@ -5,8 +5,8 @@ use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use starknet_api::core::{ContractAddress, Nonce};
 use starknet_api::data_availability::DataAvailabilityMode;
 use starknet_api::execution_resources::GasVector;
+use starknet_api::transaction::fields::signed_tx_version;
 use starknet_api::transaction::{
-    signed_tx_version,
     AccountDeploymentData,
     AllResourceBounds,
     Fee,
